requirements.txt
streamlit
pandas
sqlalchemy
psycopg2
python-dotenv
